+++ jabgui/src/main/java/org/jabref/gui/shared/SharedDatabaseLoginDialogView.java
@@ -48,9 +48,6 @@ public class SharedDatabaseLoginDialogView extends BaseDialog<Void> {
     @FXML private CheckBox autosave;
     @FXML private ButtonType connectButton;
     @FXML private CheckBox useSSL;
-    @FXML private TextField fileKeystore;
-    @FXML private PasswordField passwordKeystore;
-    @FXML private Button browseKeystore;
     @FXML private TextField serverTimezone;
     @FXML private TextField jdbcUrl;
     @FXML private CheckBox expertMode;
@@ -131,16 +128,10 @@ public class SharedDatabaseLoginDialogView extends BaseDialog<Void> {
 
         useSSL.selectedProperty().bindBidirectional(viewModel.useSSLProperty());
 
-        fileKeystore.textProperty().bindBidirectional(viewModel.keyStoreProperty());
-
         expertMode.selectedProperty().bindBidirectional(viewModel.expertModeProperty());
         jdbcUrl.textProperty().bindBidirectional(viewModel.jdbcUrlProperty());
         jdbcUrl.disableProperty().bind(viewModel.expertModeProperty().not());
 
-        fileKeystore.disableProperty().bind(viewModel.useSSLProperty().not());
-        browseKeystore.disableProperty().bind(viewModel.useSSLProperty().not());
-        passwordKeystore.disableProperty().bind(viewModel.useSSLProperty().not());
-        passwordKeystore.textProperty().bindBidirectional(viewModel.keyStorePasswordProperty());
         rememberPassword.selectedProperty().bindBidirectional(viewModel.rememberPasswordProperty());
 
         // Settings a pasted URL or the last login switched on must not stay hidden
@@ -162,9 +153,6 @@ public class SharedDatabaseLoginDialogView extends BaseDialog<Void> {
 
             EasyBind.subscribe(autosave.selectedProperty(), selected ->
                     visualizer.initVisualization(viewModel.folderValidation(), folder, true));
-
-            EasyBind.subscribe(useSSL.selectedProperty(), selected ->
-                    visualizer.initVisualization(viewModel.keystoreValidation(), fileKeystore, true));
         });
     }
 
@@ -177,10 +165,5 @@ public class SharedDatabaseLoginDialogView extends BaseDialog<Void> {
     @FXML
     private void showSaveDbToFileDialog(ActionEvent event) {
         viewModel.showSaveDbToFileDialog();
-    }
-
-    @FXML
-    private void showOpenKeystoreFileDialog(ActionEvent event) {
-        viewModel.showOpenKeystoreFileDialog();
     }
 }